refactor: Returns a 204 response instead of a 401 when a client is blocked
When we return a 401 response, the Xamarin app raise an exception
which crash the process of sending logs, which create more application
logs and therefore more requests to send logs again.

I suspect that sending a response with a sucessful status code will
reduce the overall number of request from terminated clients.